Flatten the response for a single record
- Returns an object with the record rather than an object with the key and
the record as its value.
- Change item (List of Blob) to blob (Blob).
- index-entry-number gets removed.

See
https://github.com/openregister/registers-rfcs/blob/master/content/indexes-removal/index.md
