+++ src/main/java/uk/gov/register/core/Record.java
@@ -1,10 +1,5 @@
 package uk.gov.register.core;
 
-import com.fasterxml.jackson.dataformat.csv.CsvSchema;
-import uk.gov.register.views.EntryView;
-
-import java.util.*;
-
 public class Record {
     private final Entry entry;
     private final Item item;
@@ -20,16 +15,6 @@ public class Record {
 
     public Item getItem() {
         return item;
-    }
-
-    public static CsvSchema csvSchema(Iterable<String> fields) {
-        CsvSchema entrySchema = EntryView.csvSchemaWithOmittedFields(Arrays.asList("item-hash"));
-        CsvSchema.Builder schemaBuilder = entrySchema.rebuild();
-
-        for (Iterator<CsvSchema.Column> iterator = Item.csvSchema(fields).rebuild().getColumns(); iterator.hasNext();) {
-            schemaBuilder.addColumn(iterator.next().getName(), CsvSchema.ColumnType.STRING);
-        }
-        return schemaBuilder.build();
     }
 
     @Override
