+++ src/main/java/uk/gov/register/views/v2/RecordView.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonValue;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.fasterxml.jackson.dataformat.csv.CsvSchema;
 import io.dropwizard.jackson.Jackson;
 import uk.gov.register.core.Entry;
 import uk.gov.register.core.Field;
@@ -14,15 +15,13 @@ import uk.gov.register.views.CsvRepresentationView;
 import uk.gov.register.views.ItemView;
 import uk.gov.register.views.representations.CsvRepresentation;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
 public class RecordView implements CsvRepresentationView<ObjectNode> {
-    private final Map<String, FieldValue> fieldValueMap;
-    private Iterable<Field> fields;
+        private Iterable<Field> fields;
     private Record record;
     private final ItemView itemView;
     private final ObjectMapper jsonObjectMapper = Jackson.newObjectMapper();
@@ -30,7 +29,6 @@ public class RecordView implements CsvRepresentationView<ObjectNode> {
     public RecordView(final Record record, final Map<String, FieldValue> fieldValueMap, final Iterable<Field> fields) throws FieldConversionException {
         this.fields = fields;
         this.record = record;
-        this.fieldValueMap = fieldValueMap;
         this.itemView = new ItemView(record.getItem().getSha256hex(), fieldValueMap, fields);
     }
 
@@ -54,7 +52,7 @@ public class RecordView implements CsvRepresentationView<ObjectNode> {
     public JsonNode getContent() {
         final ObjectNode result = jsonObjectMapper.createObjectNode();
         result.putPOJO("blob", itemView);
-        result.put("entry-number", this.getEntry().getEntryNumber());
+        result.put("entry-number", this.getEntry().getEntryNumber().toString());
         result.put("entry-timestamp", this.getEntry().getTimestampAsISOFormat());
         result.put("key", this.getEntry().getKey());
         return result;
@@ -62,16 +60,27 @@ public class RecordView implements CsvRepresentationView<ObjectNode> {
 
     protected ObjectNode getFlatRecordJson() {
         ObjectNode result = jsonObjectMapper.convertValue(itemView, ObjectNode.class);
-        result.put("entry-number", this.getEntry().getEntryNumber());
+        result.put("entry-number", this.getEntry().getEntryNumber().toString());
         result.put("entry-timestamp", this.getEntry().getTimestampAsISOFormat());
         result.put("key", this.getEntry().getKey());
         return result;
     }
 
+    public CsvSchema csvSchema() {
+        CsvSchema.Builder schemaBuilder = new CsvSchema.Builder();
+        schemaBuilder.addColumn("entry-number", CsvSchema.ColumnType.STRING);
+        schemaBuilder.addColumn("entry-timestamp", CsvSchema.ColumnType.STRING);
+        schemaBuilder.addColumn("key", CsvSchema.ColumnType.STRING);
+
+        for (String value : getFieldNames()) {
+            schemaBuilder.addColumn(value, CsvSchema.ColumnType.STRING);
+        }
+
+        return schemaBuilder.build();
+    }
+
     @Override
     public CsvRepresentation<ObjectNode> csvRepresentation() {
-        final List<String> fieldNames = Arrays.asList("key", "entry-number", "entry-timestamp");
-        fieldNames.addAll(getFieldNames());
-        return new CsvRepresentation<>(Record.csvSchema(fieldNames), getFlatRecordJson());
+        return new CsvRepresentation<>(csvSchema(), getFlatRecordJson());
     }
 }
